more fixes for URL issues, trailing slashes in particular
It now looks like anything that should be properly 404'ing will do so. For example, localhost:8000/aboutWHATEVER used to not 404 since all urls.py cared about was the substring "about". Now only "about" and "about/" are accepted - the rest will 404.

In addition, some pages (such as calendar and pictures) had pretty nasty behavior when it came to URL resolving and loading components from media (respectively) - this now appears to be fixed